small re-org, targeted pca, and module refactor
