+++ src/rfd/risks/raw/inflation.py
