+++ src/rfd/risks/raw/interest_rate_5y.py
